feat(api): started work on #20
