--- api/src/docgen/docgen.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DocgenService } from './docgen.service';
-
-@Module({
-  providers: [DocgenService]
-})
-export class DocgenModule {}
